refactor path controlers
