+++ organize/app/src/main/java/com/organize4event/organize/ui/activities/HomeActivity.java
@@ -19,8 +19,8 @@ import com.organize4event.organize.R;
 import com.organize4event.organize.commons.AppApplication;
 import com.organize4event.organize.commons.CircleTransform;
 import com.organize4event.organize.commons.PreferencesManager;
-import com.organize4event.organize.controllers.FirstAccessControler;
-import com.organize4event.organize.controllers.NotificationControler;
+import com.organize4event.organize.controlers.FirstAccessControler;
+import com.organize4event.organize.controlers.NotificationControler;
 import com.organize4event.organize.listeners.ControllResponseListener;
 import com.organize4event.organize.listeners.ToolbarListener;
 import com.organize4event.organize.models.FirstAccess;
