+++ organize/app/src/main/java/com/organize4event/organize/ui/activities/LoginActivity.java
@@ -31,9 +31,9 @@ import com.organize4event.organize.R;
 import com.organize4event.organize.commons.AppApplication;
 import com.organize4event.organize.commons.CustomValidate;
 import com.organize4event.organize.commons.PreferencesManager;
-import com.organize4event.organize.controllers.FirstAccessControler;
-import com.organize4event.organize.controllers.TokenControler;
-import com.organize4event.organize.controllers.UserControler;
+import com.organize4event.organize.controlers.FirstAccessControler;
+import com.organize4event.organize.controlers.TokenControler;
+import com.organize4event.organize.controlers.UserControler;
 import com.organize4event.organize.enuns.AccessPlatformEnum;
 import com.organize4event.organize.enuns.DialogTypeEnum;
 import com.organize4event.organize.enuns.LoginTypeEnum;
