+++ organize/app/src/main/java/com/organize4event/organize/ui/activities/PrivacyActivity.java
@@ -13,8 +13,8 @@ import com.google.firebase.analytics.FirebaseAnalytics;
 import com.organize4event.organize.R;
 import com.organize4event.organize.commons.AppApplication;
 import com.organize4event.organize.commons.PreferencesManager;
-import com.organize4event.organize.controllers.PrivacyControler;
-import com.organize4event.organize.controllers.UserControler;
+import com.organize4event.organize.controlers.PrivacyControler;
+import com.organize4event.organize.controlers.UserControler;
 import com.organize4event.organize.enuns.DialogTypeEnum;
 import com.organize4event.organize.enuns.PrivacyEnum;
 import com.organize4event.organize.listeners.ControllResponseListener;
