+++ organize/app/src/main/java/com/organize4event/organize/ui/activities/TermUseActivity.java
@@ -10,7 +10,7 @@ import android.widget.TextView;
 import com.google.firebase.analytics.FirebaseAnalytics;
 import com.organize4event.organize.R;
 import com.organize4event.organize.commons.PreferencesManager;
-import com.organize4event.organize.controllers.TermUseControler;
+import com.organize4event.organize.controlers.TermUseControler;
 import com.organize4event.organize.listeners.ControllResponseListener;
 import com.organize4event.organize.listeners.ToolbarListener;
 import com.organize4event.organize.models.FirstAccess;
