+++ organize/app/src/main/java/com/organize4event/organize/ui/activities/UserRegisterActivity.java
@@ -32,11 +32,11 @@ import com.mobsandgeeks.saripaar.annotation.Password;
 import com.organize4event.organize.R;
 import com.organize4event.organize.commons.CircleTransform;
 import com.organize4event.organize.commons.Mask;
-import com.organize4event.organize.controllers.FirstAccessControler;
-import com.organize4event.organize.controllers.PrivacyControler;
-import com.organize4event.organize.controllers.SettingsControler;
-import com.organize4event.organize.controllers.TermUseControler;
-import com.organize4event.organize.controllers.UserControler;
+import com.organize4event.organize.controlers.FirstAccessControler;
+import com.organize4event.organize.controlers.PrivacyControler;
+import com.organize4event.organize.controlers.SettingsControler;
+import com.organize4event.organize.controlers.TermUseControler;
+import com.organize4event.organize.controlers.UserControler;
 import com.organize4event.organize.enuns.DialogTypeEnum;
 import com.organize4event.organize.enuns.PrivacyEnum;
 import com.organize4event.organize.enuns.UserTypeEnum;
