+++ organize/app/src/main/java/com/organize4event/organize/ui/fragments/InstitutionalFragment.java
@@ -17,8 +17,8 @@ import com.afollestad.materialdialogs.MaterialDialog;
 import com.google.firebase.analytics.FirebaseAnalytics;
 import com.organize4event.organize.R;
 import com.organize4event.organize.commons.AppApplication;
-import com.organize4event.organize.controllers.InstitutionalControler;
-import com.organize4event.organize.controllers.PlanControler;
+import com.organize4event.organize.controlers.InstitutionalControler;
+import com.organize4event.organize.controlers.PlanControler;
 import com.organize4event.organize.enuns.ContactTypeEnum;
 import com.organize4event.organize.enuns.DialogTypeEnum;
 import com.organize4event.organize.enuns.PlanEnum;
