+++ organize/app/src/main/java/com/organize4event/organize/ui/fragments/SettingsFragment.java
@@ -16,8 +16,8 @@ import com.afollestad.materialdialogs.MaterialDialog;
 import com.google.firebase.analytics.FirebaseAnalytics;
 import com.organize4event.organize.R;
 import com.organize4event.organize.commons.AppApplication;
-import com.organize4event.organize.controllers.PlanControler;
-import com.organize4event.organize.controllers.SettingsControler;
+import com.organize4event.organize.controlers.PlanControler;
+import com.organize4event.organize.controlers.SettingsControler;
 import com.organize4event.organize.enuns.SettingsEnum;
 import com.organize4event.organize.listeners.ControllResponseListener;
 import com.organize4event.organize.listeners.MultipleRecyclerViewListener;
